feat(users): update router
