+++ src/api-docs/openAPIDocumentGenerator.ts
@@ -1,7 +1,7 @@
 import { OpenAPIRegistry, OpenApiGeneratorV3 } from "@asteasolutions/zod-to-openapi";
 
 import { healthCheckRegistry } from "@/api/healthCheck/healthCheckRouter";
-import { userRegistry } from "@/api/user/user.openapi.registry";
+import { userRegistry } from "@/api/user/user.openapi";
 
 export function generateOpenAPIDocument() {
 	const registry = new OpenAPIRegistry([healthCheckRegistry, userRegistry]);
