+++ src/api/user/user.openapi.ts
@@ -1,0 +1,24 @@
+import { OpenAPIRegistry } from "@asteasolutions/zod-to-openapi";
+import { z } from "zod";
+
+import { createApiResponse } from "@/api-docs/openAPIResponseBuilders";
+import { GetUserRequestSchema, UserSchema } from "@/api/user/user.model";
+
+export const userRegistry = new OpenAPIRegistry();
+
+userRegistry.register("User", UserSchema);
+
+userRegistry.registerPath({
+	method: "get",
+	path: "/users",
+	tags: ["User"],
+	responses: createApiResponse(z.array(UserSchema), "Success"),
+});
+
+userRegistry.registerPath({
+	method: "get",
+	path: "/users/{id}",
+	tags: ["User"],
+	request: { params: GetUserRequestSchema.shape.params },
+	responses: createApiResponse(UserSchema, "Success"),
+});
